patches/2.6.35: added initramfs hack
This patch simply breaks the loop that uncompresses cpio.gz
archives into initramfs. The default behaviour is to error out
when no more valid data is there in the "initrd=start,size" area.
Since we don't know in advance the exact size of the compressed
initramfs when the kernel is being compiled, the command line in the
WR switch shas 16M ("initrd=0x23000000,16777216") as a safe size.

This patch prevents a panic and allos the system to boot with a
smaller-than-expected cpio archive.