wbgen-regs: new input wb files and output headers
This is the new input (*.wb) and output (*.h) files for
device registers. I was making a huge mess with the copying
and filtering, so I decided to make order. The previous
commit adds the engine (Makefile), and the still previous removed
stuff I committed by error. Now this part is stable I hope.