build: use a unique and future-proof gateware name
The file I uploaded to ohwr (wrs3-gw-v3.0-20120801.tar.gz) is not
expected to be renamed or overwritten in the future, so this commit
and the following ones will always build properly.

We had this problem with the previous file name being replaced, so
older commits didn't build any more for lack of the required gateware
file or md5sum errors.