added kconfig infrastructure, from wrpc-sw
This is the combination of the following commits (oldest on top)
from wrpc-sw:

  89cdfcd added kconfig infrastructure, the trivial way
  5fda0c8 kconfig: use pathnames from Linux-2.6.34
  ab30804 kconfig: use 'Kconfig' as file name, not the bst-kconfig ones
  7c78880 kconfig: trivially write "WRPC" in place of "Linux Kernel"
  2c30cb4 kconfig: trivially fix most 'kernel' words in messages
  70eb3ba kconfig: add a configs/ directory and use it
  64cc10f scripts/kconfig: don't complain for default in choices

After picking them, I replaced all "WRPC" with "PPSi", and I removed
the configs/  files (unrelated to this project).

In practice, at this point we have the Kconfig infrastructure of
Linux-2.6.34, which I picked from the "bst-kconfig" package, which did
the extraction work.

One day I forward-ported this from v2.6.34 to v3.4, by applying a few
hundred kernel patches. While the final code was working, we decided
the port was not needed for wrpc-sw. Here in PPSi I keep the same version
we have in wrpc-sw.

As of this commit, the code is not used; the default configuration
file and Makefile targets are added in the next commits.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>