build: new configs/buildroot/ directory; use Kconfig to ask for cfg file
Both the buildroot and busybox config files are not hosted in
configs/buildroot/ within wr-switch-sw, where I expect to place
several such configuration files, at least to try different setups.

The name of the buildroot configuration file is not chosen using
Kconfig (but the default is silently applied.  If the user
provides a relative pathname, it is expanded to
wr-switch-sw/configs/buildroot .

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>