softpll-unify: header changes (no effect on rt_cpu.bin)
The commit is part of the effort in unifying softpll with wrpc-sw,
and later remove the duplicated code here in wr-switch-sw.

Here I copy includes/std/ from wrpc-sw, and change hw includes
to be the same as in wrpc-sw.  I'm aware rt/include/stdint.h is
better than the new rt/include/std/stdint.h but that's it for now.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>