softpll-unify: add prototypes (this changes the binary)
The commit is part of the effort in unifying softpll with wrpc-sw,
and later remove the duplicated code here in wr-switch-sw.

The addition of missing prototypes (now zeroing the diff for
softpll_ng.h between here and wrpc-sw) changes the binary being built,
but I assume we can feel safe about it.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>