softpll-unify: rename target/ to arch/ (no effect on rt_cpu.bin)
The commit is part of the effort in unifying softpll with wrpc-sw,
and later remove the duplicated code here in wr-switch-sw.

This matches commit "5d3fc51 compilation fixes" in wrpc-sw. At this
point the two arch/lm32 only differ in the linker script, that has an
mbox section in wr-switch-sw, missing in wrpc-sw.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>