softpll-unify: linker script changes (no effect on rt_cpu.bin)
The commit is part of the effort in unifying softpll with wrpc-sw,
and later remove the duplicated code here in wr-switch-sw.

This applies the effect of the following wrpc-sw commits, limited
to arch/lm32.

   583ab27 build: move libs from ldscript to LDFLAGS
   a502f9d arch/lm32/ram.ld: provide mprintf if missing
   394d140 general: add copyright notes

Additionally, it copies over crt0.S and debug.S from wrpc-sw;
but the only change is white space.

Now the only remaining difference in arch/ is the linker script.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>