softpll-unify: rewritten linker script (no effect on rt_cpu.bin)
The commit is part of the effort in unifying softpll with wrpc-sw,
and later remove the duplicated code here in wr-switch-sw.

The commit replaces the linker script with the one in wrpc-sw, with
the only addition of the "mbox" section, which is still missing from
that script.

The only remaining differences are the configurable ram and stack sizes.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>