softpll-unify: use uart files from wrpc-sw
The commit is part of the effort in unifying softpll with wrpc-sw,
and later remove the duplicated code here in wr-switch-sw.

The uart module in wrpc-sw is a superset of this one (it has vuart), but
all the registers are compatible.

The files were slightly different, but nothing substantial, mainly
because wrpc-sw evolved a little from its original form. By copying
those uart.[ch] we shorten the difference with no techinical change, even
if the binary file differs from the previous commit.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>