build: always re-wrap rootfs
We should always run the last step: if we failed in building userspace,
for example, wrapping succeeded but we want it to be done again after
fixing the userspace error.

Re-wrapping when nothing has been rebuilt costs 6 seconds or so, and
is acceptable even if nothing changed since last time (and maybe the
developer changed the override directory).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>