config: allow specifying a DNS server and default domain
If static IP addressing is used (e.g. using wrboot), it makes sense
to provide a static DNS nameserver.  This option may be useful to
people building a single image to install several switches.

If the configuration string is empty (default) nothing changes (and
/etc/resolv.conf is a symlink to /tmp/resolv.conf, so a dhcp client
can work withint a read-only filesystem).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>