softpll-unify: use rt_cpu.bin built from wrpc-sw
By now, code here and in wrpc-sw is the same. The binary
built there is working, so let's pick that one. It
can be rebuilt in wrpc-sw by running

  make wr_switch_defconfig

The exact commit used to build what you have in binaries/ (or
in /wr/lib/firmware/rt_cpu.bin in the target system) can be
easily retrieved by extracting the commit name. For this binary:

   morgana% strings binaries/rt_cpu.bin | grep wrpc
   wrpc-v2.1-for-wrnic-19-g566aa53

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>