flashing: new procedure
This changes the flashing procedure, because the previous one did not
work any more now that the default filesystem size is bigger.

With this commit no jffs2 images is generated any more, and the
filesystem image is not included as a payload in the initramfs booted
using the USB tools.

Rather, the initramfs downloads wrs-image.tar.gz using TFTP, from the
IP address that replied to DHCP, and untars it to the NAND partitions.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>