patches/buildroot: to build under ubuntu-14.04 avoid building gcc docs
Somebody decided that suboptimal formatting is now an error. So the
new makeinfo errors out when building gcc-4.3.6 because of
documentation details.  Idiocy has no limits, and we are are bound to
suffer for that.  I remember when "gnu's not unix, we have higher standards".

This patch prevents gcc documentation to be built when buildroot prepares
the cross compiler.  Bah.

/alessandro who wasted a whole day on that