patches/buildroot: to build under ubuntu-14.04, change autoconf docs
Again, suboptimal formatting is now an error in new makeinfo. Worse,
this looks correct, but the complain is there anyways.  Hack a fix,
killing a newline that make the parser report the error.

Hope we are done: nobody outside gnu is fool enough to use makeinfo.
Me excluded, obviously (see wr-switch-sw/doc).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>