userspace: fix wrsw_mrp, wrsw_mvrp, wrsw_snmpd and build them
This adds the three subdirecotries to the build, and fixes missing
includes and libs by referring to ../ as needed.  Now it looks like
everything builds.

After building, the following files remain in the source tree:

    patches/buildroot/0001-netsnmp-config-to-reduce-footprint.patch
    userspace/rootfs_override/etc/snmp/
    userspace/rootfs_override/usr/

The first one is likely needed for building builroot, but this commit
doesn't do anything about it.  The other directories are stuff that should
be installed: etc/snmp/snmpd.conf and usr/lib/{libieee8021{,q}bridge.so

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>