partitions: compatible change, and update docs
This changes partitions, mainly adding a kernel-visible mapping of
what we ha in dataflash.  NAND changes are compatible, as only empty
space is used differently.

This also includes a new barebox.bin, that you can just write
to /dev/mtd3 after booting the new kernel.

Documentation is updated to reflect these partitions, but there also
old leftovers that are fixed in this same commit.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>