userspace: new boot/install procedure for V4 partitions
This brings everything into /etc/init.d/wrs-boot-procedure that is
executed as a "sysinit" task, before anything else.  The procedure
runs from the initrd/initramfs and is concerned with the complete
update of the device, if an update is pending.

See documentation (next commit) for details about partitions and boot
procedure.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>