build: final fixes for new pathnames and procedures
The release tar file (wrs-firmware.tar) is not compressed any more, so
fix the name all around, and the extracting command too.

Similarly, the official release is a .tar file, and it is generated by
the "wraprootfs" step, so this removes "--pack" as a build option (it
was a duplicate of wrapping it up anyways).

Release files are already the ones used for installation, so flash-wrs
doesn't need to build the cpio file any more.

The "macigstring" for installation is concatenated to the cpio file by
tellling the usb-loader to copy it to RAM, so no temporary cpio archive
is created any more: release files are always used unchanged.

Wrapping rootfs creates the release tar dire
"--pack" i
Signed-off-by: Alessandro Rubini <rubini@gnudd.com>