update ppsi and fix scripts accordingly
This picks a new ppsi, which is needed for several reasons. We have
been using it for a while now, so it's a safe upgrade.  Being ppsi a
Kconfig thing itself, we now need to make wrs_defconfig first.

Also, the new ppsi.h defines INST which is pre-defined by net-snmp, so
include ieee1588_types.h alone, as we only need ClockIdentity by now.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>