snmp library: fill the ppsi information to be exported
This fills the information by parsing the output of "wr_mon -g"
and "wr_mon -p".

Parsing is data-struct driven, and the strings must match the one
printed by wr_mon. If they don't a log message is generated:

   18:02:54 192 snmpd[1468]: wrs_ppsi_parse_line: can't find key "seource:"

To test, you can export WRS_SNMP_MON_FNAME and WRS_SNMP_MON_FNAME_P to
name two text files that can be opened instead of a pipe from wr_mon.

A problem with COUNTER64 remains, but it is a bug in netsnmp. See next
commit.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>