bugfix (no effect): wr_date.conf name
Every other place calls the file wr_date.conf.

userspace/tools/wr_date.c will open the config file (it does not,
currently) but the name is different.

So we fix the executable and leave the rest as it is, since it works:
documentation matches S70wr_date and use of .config for build-time settings.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>