bugfix: kernel: use more relaxed nand timings
We used the timing from the g45ek, but ours is slower. Actually, ours
is slower only in 16-bit mode, and has the same timing when used in
8-bit mode.

Our timing setup was outside of the allowed specification, and resulted
in random and unreproducible failures.  The "visible" effect is
that reading 100MB takes 13.4s instead of 11.4s, measured with:

  time dd bs=1k count=102400 < /dev/mtd1 > /dev/null

You can try another, still slower, setup just by hand, like this
(in this order):

   devmem 0xffffe83c w 0x00051003
   devmem 0xffffe838 w 0x000f000f
   devmem 0xffffe830 w 0x04060406

but the current settings (41003, 000c000c, 02040204) are alread within
specification.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>