rootfs_override: add /usr/bin/nm and its libbfd
This is used by the next commit, where I read the symbol table of an ELF
file the lazy way.

I built the two binaries by rebuilding buildroot with

    BR2_PACKAGE_BINUTILS=y
    BR2_PACKAGE_BINUTILS_TARGET=y

but I won't add the whole binutils binary-set to the WR switch, so the
configuration items are still unset in the default configuration for
WR switch.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>