kernel: add new DF AT45DB641E setting into kernel for SCB v3.4
Standard JEDEC ID is only 24bits to identify a DF chip.
It also has an optional Extended Device Info (EDI) on bytes 4 and/or 5
that need to be read to differentiate some DF chips. (i.e, the
difference between AT45DB641E and AT45DB642D is made by byte 4).