sdb hardware information: the data
The directory hwinfo-sdb/ includes the "source", trivially written by
hand. The "hwinfo" file is just a bunch of 0xff bytes. The final
file will be a tagged format.

binaries/sdb-for-dataflash is the one that is going to be edited
at build/install time -- in theory only by the manufacturer.
I generated it running the command

   sdbfs/userspace/gensdbfs hwinfo-sdb binaries/sdb-for-dataflash

from within a checkout of fpga-config-space (another ohwr project, the
SDB home).

The default mac addresses match what we used so far: 02:34:56:78:9A:BC
and 02:34:56:78:9A:00 for the WR ports.  The file names ("eth0.ethaddr")
are there so to use a generic setting approach in barebox.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>