general: rename wrsw_pstats, wrsw_vlans, wrsw_version to be "wrs_"
Inconsistent naming has always been a problem for me. Since other
developers agree, we chose to rename all stuff to be wrs_ or wr_
for switch-specific stuff and generic WR functionality.  Ok, the
split is not very clear-cut, I agree.

This commit only fixes the tools and their users:
     wrs_version
     wrs_vlan
     wrs_pstats

It doesn't touch wrsw_hal and wrsw_rtud (nor it adds any prefix to
non-prefixed tools like load-lm32 and so on).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>