userspace: some clean-up of wrs-boot-procedure
I was bored by repeated file names, so this uses WRS_KER and other
variables to host the "magic" name.

Similarly, the use of "-old" was inconsistent, and this uses the
same pattern for all files. Actually, it should be a function
to avoid repetitions. But I'm lazy and it is not.

The upgrade we may do now is by placing files in /update, with
magic names:

    - wrs-firmware.tar: is *must* include wrs-usr.tar.gz, because
      if this file exists /usr is recreated from scratch.

         - it may include zImage, wrs-initramfs.gz, barebox.bin
	 (this happens in released firmware files, so the boot loader
	 is being written twice, until we get rid of the sam-ba tools)

    - zImage: if present, the kernel is replaced

    - wrs-initramfs.gz: if present, the initramfs is replaced

    - barebox.bin: if presente, the boot loader is replaced

This means, in practice, that you may change the whole system or one
part at a time.  I suppose only we developers will use the "one part
at a time".

But the most important change is that barebox.bin is being replaced
during boot.  In v4.0 we didn't expect to ever replace barebox, but
then we found we needed it -- so let's provide for further updates.
OTOH at91boot is not replaced, and it is not expected to, ever.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>