boot procedure: use non-blocking write to /dev/ttyGS0
If the usb cable is disconnected, any write to /dev/ttyGS0 goes to a
buffer until it fills. Then, any further write will block. This means
the messages towards the user ("uncompressing filesystem... done" etc)
caused the update procedure to stop.

This fixes the problem by using non-blocking write for the messages:
if nobody`s reading the messages are just discarded by the kernel.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>