userspace: wr_date can now set system time from WR time
This is a short-term fix to actual user needs. We'll soon
have a kernel module to track WR time, but most likely people
won't use that module until the next release.

Meanwhile, the normal "wr_date get" is enhanced to report both UTC and
TAI.

This wr_date is a simple plug-in replacement for installed switches.

Example, after setting a wrong host time on purpose:
(the example above includes extra blank lines to separate the output
from the various commands)

   wrs-192.168.16.242#wr_date get; date; wr_date get tohost; date

   1416482191.851373280 TAI
   2014-11-20 11:16:31.851373280 TAI
   2014-11-20 11:15:56.851373280 UTC

   Mon Nov 10 09:08:14 UTC 2014

   1416482191.895071872 TAI
   2014-11-20 11:16:31.895071872 TAI
   2014-11-20 11:15:56.895071872 UTC

   Thu Nov 20 11:15:56 UTC 2014

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>