userspace/wrsw_hal: reindent, fit in 80 columns: no technical change
This is a semi-automatic pass. To work on this I need to have it
in 80 columns, and sanely indented. This pass is not changing
brace placement or whatever: it is just a "fill-region" in emacs plus
manual split of strings and shortening of comments. You can
"git diff --word-diff=color" to check.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>