userspace/wrsw_hal: a Lindent pass: no technical change
I didn't take all suggestions by Lindent, I avoided to commit
the stupid ones (when on separate hunks).

This mainly changes the spacing of "if( condition)", spaces after
commas in arguments, the placing of the pointer symbol and
the placement of braces.  I would have avoided the last item, but
it's not simple.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>