userspace/include/fpga_io.h: bugfix in #define
pps_gen.c had a missing semicolon because of this bug. Fix the
macro to behave properly withing if/else. And fix the user, adding
the missing semicolon.

We all know that these two functions should be static inline, and the
fpga base pointer should be void *, but that's for later.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>