userspace/libwr: remove unused functions
shw_malloc and shw_free are not used (well, i2c_free used shw_free,
but there is no associated alloc anywhere, so I removed it too).

The idea of a "malloc-or-exit" kind of function is good, but we'd
better make sense of things (and simplify where unused legacy remains)
before finding and fixing all unlikely error paths.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>