usersspace/hal: massive (manual) rename within hal_ports.c
After this commit, all symbols in hal_ports are of the form hal_port_*
(all but one, that is not used and I'll remove it in another commit).
I'm not smart enough to track the origin of the various names, so I
need this.  I need this especially in this file because I'm going to
export stuff using shmem, to avoid a lot of context switches.

Thus, the "private" structure hal_port_state is not a typedef any
more, as a first step in becoming public. It doesn't make sense to
have two almost-identical structures, where one is copied to the other
one item at a time to deal with the minor differences. (It happened
for some reason: I'm not complaining hard, but it's time to get rid of
this).

Unfortunately, two functions were called "halexp", and relied on a
prototype shared by the RPC server (this file) and the RPC client.
While the convention made sense at first, it now shows its limits, so
the exported one is called hal_port_* like everything in the file, but
for this I need to add structure names (but I left the typedef at this
point, to avoid massive name changes around).

If I didn't introduce bugs, the code is the same as it was. I did it
one symbol at a time while build-checking each steps (and I even have
the commits to track any error).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>