userspace: hal and libwr: change prototypes for exports
This has no effect on the code functionality.

Even if it affects libwr, this is a hal-only change.  The functions
in libwr that the hal itself uses, now receive a pointer to ports[],
which is thus back a static pointer within wrsw_hal/hal_ports.c.

This change allows moving the RPC queries to be shared memory lookups,
by using the ports[] pointer in the client; so both the client and
the server can access the same shared mamory using libwr functions.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>