boot scripts: better examples
This updates binaries/wrboot-* to be more commented and complete.
It also adds wrboot-install and wrboot-nand, that I personally used.

As a side effect, "mem=64m" is removed by all command lines, as it has
been properly autodetected for a few years now, so it is not needed
any more.

Documentation is updated too.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>