userspace: remove at91 headers
The files in userspace/include/at91 were not used.

Some were straight copies for linux headers, so including
<mach/at91_pio.h> instead of <at91/at91_pio.h> makes our copy
irrelevant.  This means both the hal and rtud must include from
$LINUX, but this was already a requirement for other userspace stuff.
The Makefiles are now fixed to do that.

at91sam9g45.h was only used for AT91C_BASE_SYS, which is no more.
This is trivially 4k-before-the-end, to mmap the end page.  We were
already defining AT91C_BASE_SYS_RAW to the exact same value (only,
AT91C_BASE_SYS was a pointer to structure), so the header could go as
well.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>