userspace: clean up sub-makefiles
Long ago we used to refer to installed shared libraries and headers,
during further steps of user-space compilation.  We use static libraries
since long ago, and we have all -I pointing to ../ within the source
tree.  Get rid of unneeded installation of headers and extra -I
in the command lines.

Meanwhile, make all sub-makefile a little more similar one another
(this is trivial, mostly a side effect of the above).

The commit has no effect on build commands and results.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>