userspace/ppsi: pick new version that exports using shared memory
This version exports all status information in shared memory. This
allows a simplified wr_mon, and some diagnostics.  This is version 1
of the data structures, but we'll soon change them after some audit
to remove redundant, unused and misleading data items (or add the servo
structures, that are still missing).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>