www: bugfix: use local copy of jquery.js
Our web pages use the "jquery" javascript library, from
code.jquery.com.  With this commit we refer to a local copy of the
file, instead of always referring to the upstream instance -- which
failed when the WRS was not connected to the outside world, and
generated extra traffic anyways.

Now head.php refers to the local copy; the WRS filesystem includes both
the source form (http://code.jquery.com/jquery-1.10.1.js)
and the minimized form, unreadable but shorter, which is actually
served to the browser (http://code.jquery.com/jquery-1.10.1.min.js).

Note: commit author is benoit but then alessandro added the unminimized
javascript and wrote the commit message: any error or imprecision is a's,
not b's.