userspace/wrsw_hal: use new dot-config instead of wrsw_hal.conf
This commit uses dot-config (previous commit), in exactly the same
way as previous code. No change in behaviour is there, as confirmed
by looking at shared memory values.

As a side effect, the ports as listed in shared memory
(wrs_dump_shmem) are in the right order instead of random order
(this was due to use of lua's hash functions).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>