userspace: lib and hal: use new dot-config instead of sfp-database.conf
This commit uses dot-config (previous commit), in exactly the same way
as previous code. No change in behaviour is there, as confirmed by
looking at shared memory values.

However, the alpha value is not part of the SFP definition any more,
but it is part of the fiber type (later commits).  So it is currently
hardwired in the code, matching previous config values.

The data structures are slightly changed, so the version in shared memory
is increased too.  wrs_dump_shmem is updated at the same time
and ppsi is updated to get the new headers.

As a side effect, this uses calloc() instead of malloc(), and sets a
sensible (though unused) value in the "flags" field of the sfp data.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>