userspace: bugfix in apply-dot-config
The shell we have in the switch, unfortunately, reports variables
in "set" as quoted.  This means we had all our substitutions wrong
in snmpd.conf. This commit removes the quotes, and stuff now works better.

I got this in my syslog, for example:

   /wr/etc/snmpd.conf: line 10: Error: cannot create trapsink: '192.168.16.1'
   getaddrinfo: 'morgana' Name or service not known

Because it had to be "morgana" not "'morgana'", same for IP address.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>