userspace/hal: bugfix: mark a port in-use after error checking
bug introduced in c45f9161b9aa2aa95d0f088bce6d5af005f4ba74.

The new check for dotconfig params was added after setting
port->in_use to 1.  Thus, at each loop, the hal was calling

   ioctl(6, SIOCGIFFLAGS, {ifr_name="", ???}) = -1 ENODEV (No such device)

but that ENODEV was after a call to modprobe for the "" interface.
Actually this command:

    /sbin/modprobe -q -- netdev-

By means of /sbin/modprobe being the string in /proc/sys/kernel/modprobe .
But modprobe is busybox, so over NFS we had a zillion stat() accesses
for ld.so, uclibc.so and busybox.

Thanks Adam for noting the NFS traffic increase at that commit.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>