userspace/hal: reorder the main loop, no actual change
Every two months I've been wondering what this 100%-cpu main loop
was, to later acknowledge it was not really 100%.

This commit reorders items and makes it clear where the delay is.
I tried raising it to 100ms from 25ms, but the cpu load of wrsw_hal
didn't change, because most of the time we return to the main
loop before the timeout, due to an RPC call being served.

In my opinion we should add a check in the main loop, to call
port_update_all only 10 times per second, to actually
lower the CPU load of this process.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>