userspace/ppsi: use new master, fixing a buglet
When two free-running switches were turned on at the same time, the
slave's ppsi exited with a fatal error because it was setting system
time to a negative UTC value (small - 35).  We now don't set system
time to anything older than today -- if WR is that wrong, anything
else is better for sure.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>
Acked-by: Adam Wujek <adam.wujek@cern.ch>