doc: update the path to edit interface (edit with nano)
since v4.1 the /etc/ path is overriden by /usr/etc during booting
procedure, so the user might apply the changes only in the second folder
to have them ready at reboot.

We have also changed the editor to nano because it is more easier for
users that are not familiar with vim.