userspace/snmpd: improve counting userspace deamons in wrsBootStatusGroup
There was a problem with counting multiple instaces of a program especially
for dropbear where new instance is spawned for each ssh session.
It is impossible to predict number of instances at compile time or even in
run time. What we know is that at least one dropbear process should be running.

New method is capable of reporting when there are too many or too few processes
running (exp value is positive)or reporting when too few processes are running
(exp value is negative).

Use full path for snmpd binary.

Signed-off-by: Adam Wujek <adam.wujek@cern.ch>