www: aux clock configuration added to data structure and main menu
The configuration of the aux. clock is possible from dotconfig and so is
it from the web interface. For this purpose a new structure has been
created in wrs-data.php and a new php page (auxclk.php) has been added
to the webserver.
In addition, there is a new entry in functions.php for the help message.