rootfs: try to use IP from old dot-config when no IP available
This commit is to manage with situation that there is no IP address at the
moment of getting new dot-config. At this point IP can be available from
barebox, but not always. New functionality checks for IP configuration in the
old dot-config. IP will be reconfigured after getting new dot-config.

Signed-off-by: Adam Wujek <adam.wujek@cern.ch>