userspace/rtud: remove adding local macs
Removed code used to be to add rtu entries for local ports. Such setup suppouse
to allow access to ARM CPU from switch's ports. However, hw_addr in hal_shmem
was not filled and it end up that address 00:00:00:00:00:00 was added.

Removed solution didn't work neither we wanted it to be there.

Signed-off-by: Adam Wujek <adam.wujek@cern.ch>