www: wrs mode and ppsi configuration & checking dotconfig change by others
WR-PPSi now supports GrandMaster, Boundary Clock and FreeRunning modes
using dotconfig.
NTP server is now in PPSi configuration. When users try to set a NTP
server in BC mode, there is an alert message because of the WR timing
override.
Help messages have been changed for management.php and ptp.php because
of these changes.
When the user tries to save changes and dotconfig file was modified by
other user, the data structure must be reloaded (changes are not done)