userspace: clean up lm32-vuart
lm32-vuart was destroying the tty it was run into. And ctrl-C was
not working (the initial message "press ctrl-A" was not really
visible).

This removes part of the cruft of the program. It also adds a little
delay when no bytes are there to read from FPGA to releaf a little
the load on the ARM CPU.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>