root-password: fix for corner cases
The apply_dot-config script was wrong: it wouldn't force on the
empty root password. Thus, if it is set to something non-empty
and then changed (by a network dot-config) to empty, it won't work.

This bug only revealed on nfs-root setups.

Now, it's stupid to keep two symbols (ROOT_PWD_CLEAR and ROOT_PWD_CYPHER),
and I could simplify the script. And re-test. And change docs. No thanks.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>