www: endpoint mode data structure fix
Endpoint mode was not correctly displayed in endpointmode.php and
endpointcalibration.php when dotconfig CONFIG_PORTXX_PARAMS did not
include the protocol column.
Now, the forms are correctly displayed with and without the protocol
column in dotconfig.
NOTE: we assume that the protocol column is there or not for all ports or for none of them. (not mixture)