userspace/wr_mon: show peer and PTP state
wr_mon must be seriously audited, and turned back to be ptp-based
instead of hw-based (we may have two ptp ports on the same hw port).

Meanwhile, this adds the peer's MAC address and the PTP state for each
port in "up" state, if ptp is running there. It's not perfect, but
I find it  useful.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>