userspace: new dot-config directory; allow 'make config' in target
This factorized configuration stuff out of wrap_rootfs into a build
target.  It also adds building and installing "conf" so to allow the
following commands in the target (in /wr/etc):

	  make config

	  make oldconfig     (likely after editing dot-config)

	  make defconfig

I'm unable to build menuconfig at this time, because of missing
symbols in the target's libncurses. I hope to fix this when we
bump buildroot up (thanks Adam for that!)

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>