userspace/wrsw_hal: bugfix not updating temperatures in wr_mon/HAL
There was a problem with calculating fan timeout in libwr. Problem occurs when
there was a jump back in time. Since gettimeofday was used, delta for timeout
calculation became negative.
One solution was to use monotonic clock like in main function of HAL, or use
the same timeout calculations as main function. Second solution was chosen
since it is much simpler.

--remove calculation of fan/temperature read timeout in function
  shw_update_fans
--remove function shw_pwm_update_timeout and variable fan_update_timeout from
  libwr/fan.c as not needed anymore
--remove function shw_get_tics from libwr/util.c and libwr/util.h (not needed
  anymore)
--remove SHW_FAN_UPDATETO_DEFAULT from libwr/fan.h (not needed anymore)
--increase PORT_FAN_MS_PERIOD to 250ms (trade-off 200ms of port update and 500ms
  of temperatures update)
--remove increasing t1 in HAL;s main, since t1 was overwritten anyway by t2.

Signed-off-by: Adam Wujek <adam.wujek@cern.ch>