userspace/snmpd: allow snmpd to start without HAL and PPSI
SNMPd will try to connect to HAL's and PPSI's shmem at startup. If HAL or PPSI
is not started SNMPd would continue instead of exiting. Later when data from
shmem is needed SNMPd will try to open shmem again. By this it would be
possible to figure out what is happening on switch when HAL, PPSI or RTUd
failed to start.

Signed-off-by: Adam Wujek <adam.wujek@cern.ch>