userspace/snmpd: split wrsTimingStatus_data_fill into several functions
There was a bug in wrsTimingStatus_data_fill that even if not all of the
wrsPtpDataTable_data_fill, wrsSpllStatus_data_fill or
wrsPortStatusTable_data_fill were updated all objects in the
wrsTimingStatusGroup were updated. As a side effect it may happen that all
prev_ values were compared with exactly the same set of data (since there was
no update in particular data). By this delta were zero and wrong value was set.

The bug may be hit when snmp requests rate is lower than snmpd's cache timeout.

Signed-off-by: Adam Wujek <adam.wujek@cern.ch>